+++ src/lib/mod.rs
@@ -1,0 +1,3 @@
+pub mod spatial_vec;
+pub mod optimizer;
+pub mod rosen;
